+++ frontend/app/src/components/organisms/DailyReport/index.tsx
@@ -1,126 +1,27 @@
-import React, { useContext, useEffect, useRef, useState } from 'react'
-
-import {
-  deleteDailyReport,
-  editDailyReport,
-  getDailyReportsList,
-  postDailyReport,
-} from '../../../services/dailyReport/dailyReport'
+import React, { useContext } from 'react'
 import { AuthContext } from '../../../contexts/AuthContext'
 import { Flex } from '../../layout/Flex'
-import { getUserProfileData } from '../../../services/userprofile/userInfo'
 import DatePicker from 'react-datepicker'
 import 'react-datepicker/dist/react-datepicker.css'
-
-export const DailyReportsList = () => {
-  const Today = new Date()
-  const [dailyReports, setDailyReports] = useState([])
-  const [userId, setUserId] = useState(false)
-  const [loading, setLoading] = useState(true)
-  const [reportDateOn, setReportDateOn] = useState(Today)
-  const [editReportDateOn, setEditReportDateOn] = useState(Today)
-  const [inputData, setInputData] = useState({
-    text: '',
-    time: '',
-  })
-  const [editedDailyReport, setEditedDailyReport] = useState(false)
-  const { currentUserId, combinedTime, setCombinedTime } =
-    useContext(AuthContext)
-  const editInputRef = useRef(null)
-
-  useEffect(() => {
-    setUserId(currentUserId)
-    console.log(currentUserId, 'current')
-  }, [currentUserId])
-
-  useEffect(() => {
-    console.log(typeof userId, 'userId')
-    if (typeof userId == 'number') {
-      getDailyReportsList(userId).then((dailyReportsData) => {
-        setDailyReports(dailyReportsData)
-        console.log(dailyReportsData)
-      })
-      setLoading(false)
-    }
-  }, [userId])
-
-  useEffect(() => {
-    if (editedDailyReport) {
-      editInputRef.current.focus()
-      console.log(editInputRef.current)
-    }
-  }, [editedDailyReport.id])
-
-  // todo combined_time → usecontextを使う
-  useEffect(() => {
-    if (typeof userId == 'number') {
-      getUserProfileData(userId).then((data) => {
-        setCombinedTime(data.user.combinedTime)
-      })
-    }
-  }, [dailyReports])
-
-  const onChangeInputData = (key, value) => {
-    setInputData({ ...inputData, [key]: value })
-    console.log(inputData, 'inpuDataaaaaa')
-  }
-
-  const onChangeEditInput = (key, value) => {
-    setEditedDailyReport({ ...editedDailyReport, [key]: value })
-  }
-  const updatedDailyReports = (updatedDailyReport) => {
-    const updatedDailyReportArray = dailyReports.map((dailyReport) => {
-      if (dailyReport.id === updatedDailyReport.id) {
-        return updatedDailyReport
-      } else {
-        return dailyReport
-      }
-    })
-    setDailyReports(sortedByDate(updatedDailyReportArray))
-  }
-
-  // todo fetch update.
-  const onEditReport = (dailyReport) => {
-    editDailyReport(userId, dailyReport, editReportDateOn).then(
-      (dailyReportsData) => {
-        console.log(dailyReportsData, 'si')
-        updatedDailyReports(dailyReportsData)
-      }
-    )
-    setEditedDailyReport(false)
-  }
-
-  const sortedByDate = (data) =>
-    data
-      .slice()
-      .sort((a, b) => new Date(b.reportDateOn) - new Date(a.reportDateOn))
-
-  const handleSubmit = (event) => {
-    event.preventDefault()
-    postDailyReport(userId, inputData, reportDateOn).then((dailyReportData) => {
-      setDailyReports(sortedByDate([...dailyReports, dailyReportData]))
-    })
-    setInputData({
-      text: '',
-      time: '',
-    })
-  }
-  // todo: fix
-  const onDeleteReport = (reportId) => {
-    const deleteAndGet = async () => {
-      await deleteDailyReport(userId, reportId)
-      getDailyReportsList(userId).then((dailyReportsData) => {
-        setDailyReports(dailyReportsData)
-      })
-    }
-    deleteAndGet()
-  }
-  // todo: useRefを使う
-  const onEditReportInput = (dailyReport) => {
-    setEditedDailyReport(dailyReport)
-    setEditReportDateOn(new Date(dailyReport.reportDateOn))
-  }
-
+export const DailyReportsList = ({
+  combinedTime,
+  dailyReports,
+  editInputRef,
+  editedDailyReport,
+  editReportDateOn,
+  handleSubmit,
+  inputData,
+  loading,
+  onChangeEditInput,
+  onChangeInputData,
+  onDeleteReport,
+  onEditReport,
+  onEditReportInput,
+  reportDateOn,
+  setEditReportDateOn,
+  setReportDateOn,
+  Today,
+}) => {
   return (
     <>
       <div>総学習時間:</div>
